+++ com.xpturn.klotho/Runtime/ECS/System/Filter.cs
@@ -18,7 +18,7 @@ namespace xpTURN.Klotho.ECS
             _denseToSparse = storage1.DenseToSparse;   // span derived once in ctor (one MemoryMarshal.Cast)
             _count = storage1.Count;
             _index = 0;
-            _guard = new FilterMutationGuard(storage1.Watch);
+            _guard = FilterMutationGuard.Create(storage1, _count);
         }
 
         public bool Next(out EntityRef entity)
@@ -62,23 +62,20 @@ namespace xpTURN.Klotho.ECS
             _storage2 = storage2;
             _entities = entities;
 
-            ComponentStorageWatch watch;
-
             // For efficiency, iterate over the smaller storage
             if (storage1.Count <= storage2.Count)
             {
                 _denseToSparse = storage1.DenseToSparse;
                 _count = storage1.Count;
-                watch = storage1.Watch;
+                _guard = FilterMutationGuard.Create(storage1, _count);
             }
             else
             {
                 _denseToSparse = storage2.DenseToSparse;
                 _count = storage2.Count;
-                watch = storage2.Watch;
+                _guard = FilterMutationGuard.Create(storage2, _count);
             }
             _index = 0;
-            _guard = new FilterMutationGuard(watch);
         }
 
         public bool Next(out EntityRef entity)
@@ -134,24 +131,23 @@ namespace xpTURN.Klotho.ECS
             // Select and iterate over the smallest storage
             int min = storage1.Count;
             _denseToSparse = storage1.DenseToSparse;
-            var watch = storage1.Watch;
+            _guard = FilterMutationGuard.Create(storage1, min);
 
             if (storage2.Count < min)
             {
                 min = storage2.Count;
                 _denseToSparse = storage2.DenseToSparse;
-                watch = storage2.Watch;
+                _guard = FilterMutationGuard.Create(storage2, min);
             }
             if (storage3.Count < min)
             {
                 min = storage3.Count;
                 _denseToSparse = storage3.DenseToSparse;
-                watch = storage3.Watch;
+                _guard = FilterMutationGuard.Create(storage3, min);
             }
 
             _count = min;
             _index = 0;
-            _guard = new FilterMutationGuard(watch);
         }
 
         public bool Next(out EntityRef entity)
@@ -211,15 +207,14 @@ namespace xpTURN.Klotho.ECS
 
             int min = storage1.Count;
             _denseToSparse = storage1.DenseToSparse;
-            var watch = storage1.Watch;
+            _guard = FilterMutationGuard.Create(storage1, min);
 
-            if (storage2.Count < min) { min = storage2.Count; _denseToSparse = storage2.DenseToSparse; watch = storage2.Watch; }
-            if (storage3.Count < min) { min = storage3.Count; _denseToSparse = storage3.DenseToSparse; watch = storage3.Watch; }
-            if (storage4.Count < min) { min = storage4.Count; _denseToSparse = storage4.DenseToSparse; watch = storage4.Watch; }
+            if (storage2.Count < min) { min = storage2.Count; _denseToSparse = storage2.DenseToSparse; _guard = FilterMutationGuard.Create(storage2, min); }
+            if (storage3.Count < min) { min = storage3.Count; _denseToSparse = storage3.DenseToSparse; _guard = FilterMutationGuard.Create(storage3, min); }
+            if (storage4.Count < min) { min = storage4.Count; _denseToSparse = storage4.DenseToSparse; _guard = FilterMutationGuard.Create(storage4, min); }
 
             _count = min;
             _index = 0;
-            _guard = new FilterMutationGuard(watch);
         }
 
         public bool Next(out EntityRef entity)
@@ -284,16 +279,15 @@ namespace xpTURN.Klotho.ECS
 
             int min = storage1.Count;
             _denseToSparse = storage1.DenseToSparse;
-            var watch = storage1.Watch;
+            _guard = FilterMutationGuard.Create(storage1, min);
 
-            if (storage2.Count < min) { min = storage2.Count; _denseToSparse = storage2.DenseToSparse; watch = storage2.Watch; }
-            if (storage3.Count < min) { min = storage3.Count; _denseToSparse = storage3.DenseToSparse; watch = storage3.Watch; }
-            if (storage4.Count < min) { min = storage4.Count; _denseToSparse = storage4.DenseToSparse; watch = storage4.Watch; }
-            if (storage5.Count < min) { min = storage5.Count; _denseToSparse = storage5.DenseToSparse; watch = storage5.Watch; }
+            if (storage2.Count < min) { min = storage2.Count; _denseToSparse = storage2.DenseToSparse; _guard = FilterMutationGuard.Create(storage2, min); }
+            if (storage3.Count < min) { min = storage3.Count; _denseToSparse = storage3.DenseToSparse; _guard = FilterMutationGuard.Create(storage3, min); }
+            if (storage4.Count < min) { min = storage4.Count; _denseToSparse = storage4.DenseToSparse; _guard = FilterMutationGuard.Create(storage4, min); }
+            if (storage5.Count < min) { min = storage5.Count; _denseToSparse = storage5.DenseToSparse; _guard = FilterMutationGuard.Create(storage5, min); }
 
             _count = min;
             _index = 0;
-            _guard = new FilterMutationGuard(watch);
         }
 
         public bool Next(out EntityRef entity)
@@ -352,7 +346,7 @@ namespace xpTURN.Klotho.ECS
             _denseToSparse = storage1.DenseToSparse;
             _count = storage1.Count;
             _index = 0;
-            _guard = new FilterMutationGuard(storage1.Watch);
+            _guard = FilterMutationGuard.Create(storage1, _count);
         }
 
         public bool Next(out EntityRef entity)
@@ -403,21 +397,19 @@ namespace xpTURN.Klotho.ECS
             _exclude = exclude;
             _entities = entities;
 
-            ComponentStorageWatch watch;
             if (storage1.Count <= storage2.Count)
             {
                 _denseToSparse = storage1.DenseToSparse;
                 _count = storage1.Count;
-                watch = storage1.Watch;
+                _guard = FilterMutationGuard.Create(storage1, _count);
             }
             else
             {
                 _denseToSparse = storage2.DenseToSparse;
                 _count = storage2.Count;
-                watch = storage2.Watch;
+                _guard = FilterMutationGuard.Create(storage2, _count);
             }
             _index = 0;
-            _guard = new FilterMutationGuard(watch);
         }
 
         public bool Next(out EntityRef entity)
@@ -475,14 +467,13 @@ namespace xpTURN.Klotho.ECS
 
             int min = storage1.Count;
             _denseToSparse = storage1.DenseToSparse;
-            var watch = storage1.Watch;
+            _guard = FilterMutationGuard.Create(storage1, min);
 
-            if (storage2.Count < min) { min = storage2.Count; _denseToSparse = storage2.DenseToSparse; watch = storage2.Watch; }
-            if (storage3.Count < min) { min = storage3.Count; _denseToSparse = storage3.DenseToSparse; watch = storage3.Watch; }
+            if (storage2.Count < min) { min = storage2.Count; _denseToSparse = storage2.DenseToSparse; _guard = FilterMutationGuard.Create(storage2, min); }
+            if (storage3.Count < min) { min = storage3.Count; _denseToSparse = storage3.DenseToSparse; _guard = FilterMutationGuard.Create(storage3, min); }
 
             _count = min;
             _index = 0;
-            _guard = new FilterMutationGuard(watch);
         }
 
         public bool Next(out EntityRef entity)
@@ -545,15 +536,14 @@ namespace xpTURN.Klotho.ECS
 
             int min = storage1.Count;
             _denseToSparse = storage1.DenseToSparse;
-            var watch = storage1.Watch;
+            _guard = FilterMutationGuard.Create(storage1, min);
 
-            if (storage2.Count < min) { min = storage2.Count; _denseToSparse = storage2.DenseToSparse; watch = storage2.Watch; }
-            if (storage3.Count < min) { min = storage3.Count; _denseToSparse = storage3.DenseToSparse; watch = storage3.Watch; }
-            if (storage4.Count < min) { min = storage4.Count; _denseToSparse = storage4.DenseToSparse; watch = storage4.Watch; }
+            if (storage2.Count < min) { min = storage2.Count; _denseToSparse = storage2.DenseToSparse; _guard = FilterMutationGuard.Create(storage2, min); }
+            if (storage3.Count < min) { min = storage3.Count; _denseToSparse = storage3.DenseToSparse; _guard = FilterMutationGuard.Create(storage3, min); }
+            if (storage4.Count < min) { min = storage4.Count; _denseToSparse = storage4.DenseToSparse; _guard = FilterMutationGuard.Create(storage4, min); }
 
             _count = min;
             _index = 0;
-            _guard = new FilterMutationGuard(watch);
         }
 
         public bool Next(out EntityRef entity)
@@ -621,16 +611,15 @@ namespace xpTURN.Klotho.ECS
 
             int min = storage1.Count;
             _denseToSparse = storage1.DenseToSparse;
-            var watch = storage1.Watch;
+            _guard = FilterMutationGuard.Create(storage1, min);
 
-            if (storage2.Count < min) { min = storage2.Count; _denseToSparse = storage2.DenseToSparse; watch = storage2.Watch; }
-            if (storage3.Count < min) { min = storage3.Count; _denseToSparse = storage3.DenseToSparse; watch = storage3.Watch; }
-            if (storage4.Count < min) { min = storage4.Count; _denseToSparse = storage4.DenseToSparse; watch = storage4.Watch; }
-            if (storage5.Count < min) { min = storage5.Count; _denseToSparse = storage5.DenseToSparse; watch = storage5.Watch; }
+            if (storage2.Count < min) { min = storage2.Count; _denseToSparse = storage2.DenseToSparse; _guard = FilterMutationGuard.Create(storage2, min); }
+            if (storage3.Count < min) { min = storage3.Count; _denseToSparse = storage3.DenseToSparse; _guard = FilterMutationGuard.Create(storage3, min); }
+            if (storage4.Count < min) { min = storage4.Count; _denseToSparse = storage4.DenseToSparse; _guard = FilterMutationGuard.Create(storage4, min); }
+            if (storage5.Count < min) { min = storage5.Count; _denseToSparse = storage5.DenseToSparse; _guard = FilterMutationGuard.Create(storage5, min); }
 
             _count = min;
             _index = 0;
-            _guard = new FilterMutationGuard(watch);
         }
 
         public bool Next(out EntityRef entity)